Refactor: import fixes
